feat: refactor time calculation and formatting in FixedBottomBar
